add seperate LRU module
