+++ simulator/CMakeLists.txt
@@ -28,6 +28,7 @@ set(CPPS
     infra/config/config.cpp
     infra/ports/ports.cpp
     infra/cache/cache_tag_array.cpp
+    infra/lru/LRUCacheInfo.cpp
     memory/memory.cpp
     memory/hierarchied_memory.cpp
     memory/plain_memory.cpp
