+++ simulator/infra/lru/LRUCacheInfo.cpp
@@ -1,0 +1,45 @@
+/**
+ * LRUCacheInfo.cpp
+ * Implementation of the least recently used line replacement module.
+ * @author Oleg Ladin, Denis Los
+ */
+
+ #include <cassert>
+
+ #include "infra/lru/LRUCacheInfo.h"
+
+ LRUCacheInfo::LRUCacheInfo( std::size_t ways)
+    : ways( ways)
+{
+    assert( ways != 0u);
+
+    lru_hash.reserve( ways);
+
+    for ( std::size_t i = 0; i < ways; i++)
+    {
+        lru_list.push_front( i);
+        lru_hash.emplace( i, lru_list.begin());
+    }
+}
+
+void LRUCacheInfo::touch( std::size_t way)
+{
+    const auto lru_it = lru_hash.find( way);
+    assert( lru_it != lru_hash.end());
+
+    // Put the way to the head of the list
+    lru_list.splice( lru_list.begin(), lru_list, lru_it->second);
+}
+
+std::size_t LRUCacheInfo::update()
+{
+    // remove the least recently used element from the tail
+    std::size_t lru_elem = lru_list.back();
+    lru_list.pop_back();
+
+    // put it to the head
+    auto ptr = lru_list.insert( lru_list.begin(), lru_elem);
+    lru_hash.insert_or_assign( lru_elem, ptr);
+
+    return lru_elem;
+}
